changed float to integer

+++ 0x01-python_async_function/0-basic_async_syntax.py
@@ -8,7 +8,7 @@ import asyncio
 import random
 
 
-async def wait_random(maxdelay:float=10)->float:
+async def wait_random(maxdelay:int=10)->float:
     """
     This function generates a random time in seconds
     and returns the value after waiting for the supposed time
